Add multiple goals
Need to call set_goal for each goal. It will add the goals IDs in a vector. During the first search, the first goal to be find will be defined as the definitive goal. If used via or_sbpl_for_ada, the heuristic is the minimal one between the heuristics to the different goals.